+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `VQT Avatar Converter Settings` component to save converter settings to the avatar.
 - Add feature to change material conversion settings for each material.
+- Add material conversion settings other than Toon Lit shader.
+    - Material Replacement
 
 ### Changed
 - Require VRChat SDK 3.3.0 or later.